requirements.txt
Django==1.11.8
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.10
djangorestframework-jwt==1.11.0
